added inference network variant
